requirements.txt
python-docx>=1.1.2
openpyxl>=3.1.5
docxtpl>=0.6.9
Jinja2>=3.1.4
pandas>=2.0.0
numpy>=1.24.0
Pillow>=10.0.0
click>=8.1.0
python-dateutil>=2.8.0
